# docs/swagger.yaml
basePath: /api
definitions:
  ds.AnalyseBooksBadgeDTO:
    properties:
      application_id:
        type: integer
      count:
        type: integer
    type: object
  ds.AnalyseBooksDTO:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      books:
        items:
          $ref: '#/definitions/ds.BookInApplicationDTO'
        type: array
      completion_date:
        type: string
      conjunction_freq:
        type: number
      creation_date:
        type: string
      creator_login:
        type: integer
      forming_date:
        type: string
      id:
        type: integer
      lexical_diversity:
        type: number
      moderator_login:
        type: integer
      responce:
        type: string
      status:
        type: integer
    type: object
  ds.AnalyseBooksResolveRequest:
    properties:
      action:
        description: '"complete" | "reject"'
        type: string
    required:
    - action
    type: object
  ds.AnalyseBooksUpdateRequest:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      conjunction_freq:
        type: number
      lexical_diversity:
        type: number
    type: object
...
  ds.BookInApplicationDTO:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      book_id:
        type: integer
      conjunction_freq:
        type: number
      description:
        type: string
      image:
        type: string
      lexical_diversity:
        type: number
      text:
        type: string
      title:
        type: string
    type: object
  ds.BookToApplicationUpdateRequest:
    properties:
      description:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления книгами, заявками и анализом текстов
  title: API для системы анализа книг
  version: "1.0"
paths:
  /analysebooks:
    get:
      description: Возвращает отфильтрованный список всех сформированных заявок (кроме
        черновиков и удаленных).
      parameters:
      - description: Фильтр по статусу заявки
        in: query
        name: status
        type: string
      - description: Фильтр по дате 'от' (формат YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Фильтр по дате 'до' (формат YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.AnalyseBooksDTO'
            type: array
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок (авторизованный пользователь)
      tags:
      - analysebooks
  /analysebooks/{id}:
    delete:
      description: Логически удаляет заявку, переводя ее в статус "удалена".
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку (авторизованный пользователь)
      tags:
      - analysebooks
    get:
      description: Возвращает полную информацию о заявке, включая привязанные книги.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.AnalyseBooksDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить одну заявку по ID (авторизованный пользователь)
      tags:
      - analysebooks
    put:
      consumes:
      - application/json
      description: Позволяет пользователю обновить поля своей заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.AnalyseBooksUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные заявки (авторизованный пользователь)
      tags:
      - analysebooks
  /analysebooks/{id}/books/{book_id}:
    delete:
      description: Удаляет связь между заявкой и книгой.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID книги
        in: path
        name: book_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить книгу из заявки (авторизованный пользователь)
      tags:
      - analysebooks-m-m
    put:
      consumes:
      - application/json
      description: Изменяет дополнительное описание для конкретной книги в рамках
        одной заявки.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID книги
        in: path
        name: book_id
        required: true
        type: integer
      - description: Новое описание
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.BookToApplicationUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить описание книги в заявке (авторизованный пользователь)
      tags:
      - analysebooks-m-m
  /analysebooks/{id}/form:
    put:
      description: Переводит заявку из статуса "черновик" в "сформирована".
      parameters:
      - description: ID заявки (черновика)
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Не все поля заполнены
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку (авторизованный пользователь)
      tags:
      - analysebooks
  /analysebooks/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Модератор завершает (с расчетом) или отклоняет заявку.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Действие: ''complete'' или ''reject'''
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/ds.AnalyseBooksResolveRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить заявку (только модератор)
      tags:
      - analysebooks
  /analysebooks/cart:
    get:
      description: Возвращает ID черновика текущего пользователя и количество книг
        в нем.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.AnalyseBooksBadgeDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить информацию для иконки корзины (авторизованный пользователь)
      tags:
      - analysebooks
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"